# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sample API",
    "description": "Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/) or HTML.",
    "version": "0.1.9"
  },
  "servers": [
    {
      "url": "",
      "description": " Main (production) server"
    }
  ],
  "paths": {
    "/api/getAllCategory": {
      "get": {
        "summary": "getting all categories",
        "description": "get all categories",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/api/getCategoryByName/{name}": {
      "get": {
        "summary": "getting specific category",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": "true",
            "description": "filtering categories by name"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/api/addCategory": {
      "post": {
        "summary": "add categories",
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "name"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/api/addUser": {
      "post": {
        "summary": "Registeration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    }
  }
}